(MoritzLangenstein) Refactored command line interface
