+++ iaso/click/validators.py
@@ -5,18 +5,25 @@ import click
 from ..curation.validator import CurationValidator
 
 
-def load_registered_validators(ctx):
-    ctx.obj["validators"] = {
-        entry_point.name: entry_point
-        for entry_point in pkg_resources.iter_entry_points("iaso.plugins")
-    }
+def ensure_registered_validators(ctx):
+    registered_validators = ctx.obj.get("validators")
+
+    if registered_validators is None:
+        registered_validators = ctx.obj["validators"] = {
+            entry_point.name: entry_point
+            for entry_point in pkg_resources.iter_entry_points("iaso.plugins")
+        }
+
+    return registered_validators
 
 
 def validate_validators(ctx, param, value):
+    registered_validators = ensure_registered_validators(ctx)
+
     validators = dict()
 
     for validator_name in set(value):
-        Validator = ctx.obj["validators"].get(validator_name, None)
+        Validator = registered_validators.get(validator_name, None)
 
         if Validator is None:
             raise click.BadParameter(
@@ -36,7 +43,7 @@ def validate_validators(ctx, param, value):
                     f"{validator_name} does not export a non-abstract subclass of 'iaso.curation.validator.CurationValidator'."
                 )
 
-            ctx.obj["validators"][validator_name] = Validator
+            registered_validators[validator_name] = Validator
 
         validators[validator_name] = Validator
 
@@ -47,7 +54,9 @@ def list_validators(ctx, param, value):
     if not value or ctx.resilient_parsing:
         return
 
-    if len(ctx.obj["validators"]) == 0:
+    registered_validators = ensure_registered_validators(ctx)
+
+    if len(registered_validators) == 0:
         click.echo(
             "No validator modules have been registered under 'iaso.plugins' yet."
         )
@@ -56,7 +65,7 @@ def list_validators(ctx, param, value):
             "The following validator modules have been registered under 'iaso.plugins':"
         )
 
-    for validator in ctx.obj["validators"]:
+    for validator in registered_validators:
         click.echo(f"- {validator}")
 
     ctx.exit()
